Classify CO2 with an option object.
This just an example of what it could look like. What I like about
this approach is that we open up the API with a configurtion object
so when in the future when we have multiple algoritms to
calculate the CO2 emission, it can always be made backward compatible
or you as a user can choose which one to use.

If you like it, I can do the same with the hosting APIs.