fix: use dataReloadRatio as multiplier instead of its complement in SWD v4
Summary

In the original formula of the SWD v4 model the data cache ratio is used
and the emissions per byte are multiplied with (1 - data cache ratio).
The CO2js implementation uses the data reload ratio = (1 - data cache
ratio). But this is not reflected in the computation, which still
multiplies by (1 - data reload ratio).

This PR fixes that.